requirements.txt
streamlit
python-dotenv
youtube-transcript-api
google-api-python-client
google-generativeai
textblob
scikit-learn
matplotlib
wordcloud
